FarmForgeController
Added FarmForgeController with some common / base functionality
