+++ StrykerDG.FarmForge.Api/Controllers/LocationController.cs
@@ -14,7 +14,7 @@ namespace StrykerDG.FarmForge.LocalApi.Controllers
 {
     [ApiController]
     [Route("Locations")]
-    public class LocationController : ControllerBase
+    public class LocationController : FarmForgeController
     {
         private IActorRef LocationActor { get; set; }
 
@@ -44,14 +44,7 @@ namespace StrykerDG.FarmForge.LocalApi.Controllers
                 )
             );
 
-            var resultType = result.GetType();
-            if(resultType == typeof(Exception))
-            {
-                var ex = (Exception)result;
-                return Ok(FarmForgeApiResponse.Failure(ex.Message));
-            }
-            else
-                return Ok(FarmForgeApiResponse.Success(result));
+            return ValidateActorResult(result);
         }
 
         [HttpPatch]
@@ -65,14 +58,7 @@ namespace StrykerDG.FarmForge.LocalApi.Controllers
                )
             );
 
-            var resultType = result.GetType();
-            if (resultType == typeof(Exception))
-            {
-                var ex = (Exception)result;
-                return Ok(FarmForgeApiResponse.Failure(ex.Message));
-            }
-            else 
-                return Ok(FarmForgeApiResponse.Success(result));
+            return ValidateActorResult(result);
         }
 
         [HttpDelete("{locationId}")]
@@ -80,15 +66,8 @@ namespace StrykerDG.FarmForge.LocalApi.Controllers
         public async Task<IActionResult> DeleteLocation(int locationId)
         {
             var result = await LocationActor.Ask(new AskToDeleteLocation(locationId));
-            
-            var resultType = result.GetType();
-            if (resultType == typeof(Exception))
-            {
-                var ex = (Exception)result;
-                return Ok(FarmForgeApiResponse.Failure(ex.Message));
-            }
-            else 
-                return Ok(FarmForgeApiResponse.Success(result));
+
+            return ValidateActorResult(result);
         }
     }
 }
